Split the settings router off from the UserProfile relationship to match architecture of the backend
Created conditional rendering system
matching jwt vs the current userData fetch on the
API to determine if the user can access
the settings/edit page of a profile.

Drilled in the state of userData to the Settings
page to reduce API calls

Began Settings.jsx, still determining
what should be editable and how to edit.