Seeded database with a large amount of genres that should apply to most usecases, based off the steam database of most relevant genretags.
Regarding this db update, the data was seeded
in alphabetical order so in the future
algorithms for searching can run at best
time complexity.

Controller logic updated for fetching games
related to a specific genre from the genres page.
When you click a genre it will now do a multi
level sort where if the game does not contain
the genre in its genre table, it leaves the sort,
and after another simple .sort(a,b) if the
queried genre is not in the first index, it will
null and be left out of the return. This way
we can include only the most relevant games to
a genre.

Further work needs to be done in this area
with regards to a weighting system where
we dont regard all activity as positive,
we will need to implement an improved
algorithm later.