Beginning work of migrating away from using gameID as a parameter in queries, instead using slug routing
Redesigned submission editing and approving
game tickets in accordance with this new design.

Slugs are generated on submission of game,
and are able to be changed if the title of the
game is updated or changed by a user ticket.

Uniqueness is ensured by appending a uniqueID
after the slug name.

Schema changes have been made, db not migrated
yet.

Looking through all routers that have any
dependency on gameID and converting into a
slug routing system.