Updated genre system seeded into the database.
Parent Genres contain unique 1 to many child
genres.

There are pseudoparent genres for tagging
characteristics about the game that transcend
genre specific attributes. This was a work around
for keeping a 1 to many and should function well
if queries are properly written in controllers.

The slug routing system worked well for games,
so the system was expanded and takes care of
genres now. The slug is generated through the same
util function.