Major Change:
Had to drop and reseed the database because
WHERE queries in prisma only function based off
@unique fields. The DB is the same except
slugs in both Game and Genre models are now
tagged with @unique after the drop/reseed/migrate

Controller function rewritten for fetching the
game contents of a genre, and its respective
subgenres if there are any that exist.

There is currently no sorting applied to how the
data is returned, which will be implemented later
with a simple sorts like max reviews, max ratings,
release date, which give the best information for
users with minimal amount of complexity in query
logic.

Seed scripts changed and added just for future
QOL.

Router updated with better naming practices.

The page for receiving the content discussed above
has been implemented in its MVP form.

Changes to Genres.jsx for feeding in params to
redir to the genres its advertising.